\name{dummies-package}
\Rdversion{1.1}
\alias{dummies-package}
\alias{dummies}
\docType{package}
\title{
  Create and manipulate dummy variables flexibly and efficiently
}

\description{
  Contains functions to create dummy variables flexibly using 
  \code{model.matrix} returning them as either matrices or data frames for
  further analysis.  Alsoi, contains methods, for manipulating dummy
  variables.
}

\details{
\tabular{ll}{
Package: \tab dummies\cr
Type: \tab Package\cr
Version: \tab 1.5.3\cr
Date: \tab 2011-03-02\cr
License: \tab LGPL-3 \cr
LazyLoad: \tab yes\cr
}
}
\author{
Author: Christopher Brown

Maintainer: Christopher Brown
}

\references{
  http://wiki.r-project.org/rwiki/doku.php?id=tips:data-manip:create_indicator

  http://blog.opendatagroup.com/2009/09/30/r-the-dummies-package
}

\keyword{ package }
\seealso{
   \code{link{dummy}}
}
\examples{
   # See examples in the documentation for dummy
}
