% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.dummy.R
\name{get.dummy}
\alias{get.dummy}
\title{Get the dummy variable columns from a data frame.}
\usage{
get.dummy(data, name = NULL)
}
\arguments{
\item{data}{A data.frame with an dummies attributes}

\item{name}{Optional.  The name of a variable.}
}
\value{
The subset of \code{data} that are dummy columns.
}
\description{
Given a data.frame and an optional variables name, return only the columns
that are dummy variables
}
\details{
This uses \code{\link{which.dummy}} to identify the dummy columns.
}
\examples{

    data( iris ) 
    d <- dummy.data.frame( iris )
    get.dummy( d, 'Species' )

}
\seealso{
\code{\link{which.dummy}}, \cr
  \code{\link{dummy.data.frame}} \cr
}

