% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.dummy.R
\name{which.dummy}
\alias{which.dummy}
\title{Identify which columns are dummy variables on a data frame.}
\usage{
which.dummy(data, name = NULL)
}
\arguments{
\item{data}{An object with a 'dummies' attribute}

\item{name}{Optional.  The name of a column that has been expanded to a
dummy variable.}
}
\value{
integer vector of column indices corresponding to the dummy
variable(s)
}
\description{
Given a data frame and an optional variable name, \code{which.dummy}
identifies which colummns are dummy variables by the column index.
}
\details{
Requires an \code{dummies} attribute, usually created by
\code{\link{dummy.data.frame}}.



Given a data frame and an optional variable name, returns the indices of the
dummy columns.
}
\examples{

  data(iris)
  dat <- dummy.data.frame(iris)
  which.dummy(dat)

}
\seealso{
\code{\link{dummy.data.frame}}, \cr
  \code{\link{dummy}} \cr
}

